requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
sqlalchemy.orm 
werkzeug.exceptions
werkzeug.security
functools